Warning fixed.

+++ LHAPDF/lhapdf5.5.1/src/getdatapath.cxx
@@ -45,7 +45,7 @@ extern "C" {
     //for (size_t i = 0; i < SIZE; ++i) {
     //  tmp[i] = ' ';
     //}
-    for (size_t i = SIZE-1; i >= 0; --i) {
+    for (int i = SIZE-1; i >= 0; --i) {
       if (tmp[i] != ' ') break;
       tmp[i] = '\0';
     }
